ci: fix build process
