+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,18 +1,5 @@
-import NextAuth, { getServerSession } from "next-auth";
-import GithubProvider from "next-auth/providers/github";
+import NextAuth from "next-auth";
+import { authOptions } from "./authOptions";
 
-export const authOptions = {
-  providers: [
-    GithubProvider({
-      clientId: process.env.GITHUB_ID ?? "",
-      clientSecret: process.env.GITHUB_SECRET ?? "",
-    }),
-  ],
-};
-
-export const handler = NextAuth(authOptions);
+const handler = NextAuth(authOptions);
 export { handler as GET, handler as POST };
-
-const getSession = () => getServerSession(authOptions)
-
-export { getSession }
